Catch error caused by lack of drive space
One error report indicated a crash when there was not enough drive space
to generate the mod, this is now caught and handled appropriately. The
error message has also been added to all of the localisation files.

Resolves #20

The PathTooLongError error message has also been fixed, as it previously
just read "TOOO LONG" followed by the permission denied message, which
kind of worked as the solution might be the same, but was not accurate
and also included the placeholder message.