Add Western Star 5700XE and fix 57X
The new Western Star 5700XE has been added, along with its templates
which are .gitignored. The 57X has also been updated with proper
support for the Aero II cabin deflector, as either I missed it the first
time, or it was added in a subsequent update.

The vehicle-codes file was also updated for the 5700XE as well as the
Feldbinder trailers, and scroll bars have been added to every truck and
trailer list except for the bus mod lists, as they're now needed.