Make half of main screen
The left half of the main screen has now been laid out, with all of the
required text entry fields, including help buttons and warning buttons
which will be used in the future. The new tab system for working with
multiple paint jobs has also started to be implemented, with dynamic
tab renaming.