Create VehicleSelections from vehicle list
VehicleSelections are now created programatically from the
VehicleDatabase list, which has also been split into the different
categories needed. All of the components have been switched to using
the control and container system rather than being positioned manually,
so that the ScrollContainer functions as needed. As such the spacing is
now a little off, and will need to be tweaked as I learn the system
better.