Catch KeyError crash when copying main .dds
When copying a vehicle's template from its template zip, KeyErrors will
now be caught, and the empty placeholder will be copied instead. This
will prevent a program crash, but if it ever happens it means something
has gone wrong, as the main_dds_name must be incorrect, or the template
zip must be missing a file.