Prevent overscrolling vehicle lists
Whenever scrolling a vehicle list via the scrollwheel, the height of the
list's canvas and frame are now checked, and if the frame is smaller
than the canvas (so the list does not overflow the window) the list will
not scroll. This prevents short lists (such as filtered results) from
being scrollable up and down, instead keeping the elements pinned to the
top of the screen, as expected.